Prices are now double and renamed rules
- Made prices double
- Renamed rules, so they're all start with `Rule`

+++ src/main/java/hu/czompi/ch_app/inventory/products/Product.java
@@ -14,5 +14,5 @@ public abstract class Product {
      * Product cost
      * @return Price of the product
      */
-    public abstract int getPrice();
+    public abstract double getPrice();
 }
